ch gyp about configurable filter commands, add "pymod_do_main" filter
The syntax for gyp commands was "<(command)", which will execute "command" in a shell. Now it's "<filter(command)", and if filter is omitted command is run in a shell. Else, it depends on "filter" what happens. 

Add the filter "pymod_do_main", which loads the python module that's the first part of "command", passes the rest of "command" to that module's DoMain() function, and then uses the result of this function. That's roughly equivalent to `<(python module.py parameters)`, but can be a lot faster for modules that have nontrivial startup time (because they are now imported only once by gyp). 

This speeds up `build/gyp_chromium` by 8s on my system (with the change to gyp_chromium in http://codereview.chromium.org/7035004). 

BUG=http://crbug.com/82230

Review URL: http://codereview.chromium.org/7034007/



git-svn-id: https://gyp.googlecode.com/svn/trunk@924 78cadc50-ecff-11dd-a971-7dbc132099af